requirements.txt
azure-functions
psycopg2